requirements.txt
fastapi
uvicorn
prometheus-client
python-dotenv
requests